Revert behavior from #463 on Windows.
See PR for discussion. Due to os specific issues, we can't fix the bug
from #463, as it causes other problems on windows

+++ ccmlib/common.py
@@ -246,7 +246,8 @@ def make_cassandra_env(install_dir, node_path, update_conf=True):
         sh_file = os.path.join(BIN_DIR, CASSANDRA_SH)
     orig = os.path.join(install_dir, sh_file)
     dst = os.path.join(node_path, sh_file)
-    shutil.copy(orig, dst)
+    if not is_win() or not os.path.exists(dst):
+        shutil.copy(orig, dst)
 
     if update_conf and not (is_win() and get_version_from_build(node_path=node_path) >= '2.1'):
         replacements = [
